# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API Documentation",
        "description": "Swagger API documentation for my Laravel project",
        "contact": {
            "email": "contact@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/countries": {
            "get": {
                "tags": [
                    "Countries"
                ],
                "summary": "Get all countries",
                "operationId": "edaabf3348bf71d9cf65b87f44ec29ce",
                "responses": {
                    "200": {
                        "description": "list of countries",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {}
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Countries"
                ],
                "summary": "Create a new country",
                "operationId": "f709f3add391d9d07db00fffe3b55c7c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Country"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Country created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Country"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/countries/{id}": {
            "get": {
                "tags": [
                    "Countries"
                ],
                "summary": "Get a single country",
                "operationId": "da4b3e218be51657a5a434a108b06d04",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Country details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Country"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Country not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Countries"
                ],
                "summary": "Update an existing country",
                "operationId": "d37aee394fdc82fe8666cd62ee5e0334",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Country"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Country updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Country"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Country not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Countries"
                ],
                "summary": "Delete a country",
                "operationId": "c18bfce8a4b6d48c70f7b00a2104c411",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Country deleted successfully"
                    },
                    "404": {
                        "description": "Country not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Country": {
                "required": [
                    "country_name",
                    "capital",
                    "population",
                    "region",
                    "currency",
                    "language",
                    "flag"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "country_name": {
                        "type": "string",
                        "example": "Morocco"
                    },
                    "capital": {
                        "type": "string",
                        "example": "Rabat"
                    },
                    "population": {
                        "type": "integer",
                        "example": 37000000
                    },
                    "region": {
                        "type": "string",
                        "example": "North Africa"
                    },
                    "currency": {
                        "type": "string",
                        "example": "MAD"
                    },
                    "language": {
                        "type": "string",
                        "example": "Arabic"
                    },
                    "flag": {
                        "type": "string",
                        "example": "https://example.com/flags/morocco.png"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Countries",
            "description": "Countries"
        }
    ]
}